+++ app/session/index.ts
@@ -1,1 +1,8 @@
 export { useBackendSession, type BackendSessionController } from "./useBackendSession";
+export {
+  useSignedInSession,
+  type AuthMode,
+  type EncryptedSyncState,
+  type SignedInSessionController,
+} from "./useSignedInSession";
+